Added a visualization for a browser tile based procedural map generator
The process picks an origin in the central third of the height and width, then builds out one tile at a time with correction processes for if it gets stuck or if it generates an undesirably small map. You can chage the mapSize number at the top of the .js document to see different sizes in the level, though I would suggest avoiding anything under 20 and anything over 100.